Create seperate audit lambda and call it from app.py, add missing activity route, refactor code
